move header filter to filter folder
