+++ packages/devextreme/js/__internal/grids/pivot_grid/field_chooser/m_field_chooser_base.ts
@@ -15,7 +15,7 @@ import {
   headerFilterMixin,
   HeaderFilterView as HeaderFilterViewBase,
   updateHeaderFilterItemSelectionState,
-} from '@ts/grids/grid_core/header_filter/m_header_filter_core';
+} from '@ts/grids/grid_core/filter/m_header_filter_core';
 import gridCoreUtils from '@ts/grids/grid_core/m_utils';
 import sortingMixin from '@ts/grids/grid_core/sorting/m_sorting_mixin';
