+++ packages/devextreme/js/__internal/grids/tree_list/module_not_extended/header_filter.ts
@@ -1,4 +1,4 @@
-import { headerFilterModule } from '@ts/grids/grid_core/header_filter/m_header_filter';
+import { headerFilterModule } from '@ts/grids/grid_core/filter/m_header_filter';
 
 import core from '../m_core';
