+++ packages/devextreme/js/ui/grid_core/ui.grid_core.header_filter.js
@@ -1,2 +1,2 @@
 // deprecated
-export * from '../../__internal/grids/grid_core/header_filter/m_header_filter';
+export * from '../../__internal/grids/grid_core/filter/m_header_filter';
